SCM - extract history item hover content creation
